pi-app/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.9.2
pillow==10.4.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
httpx==0.27.2
python-dotenv==1.0.1
rapidfuzz==3.9.6
structlog==24.4.0
cloudscraper==1.2.71
prometheus-fastapi-instrumentator==6.1.0
